requirements.txt
spotipy
python-dotenv
fastapi
uvicorn
sqlalchemy
librosa
soundfile
numpy
yt-dlp
scikit-learn

